requirements.txt
flask
flask-cors
torch
torchvision
timm
pillow
numpy